+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/switch/logo_lang/switch-comp/emfswitch/operation/EqualsOperation.java
@@ -1,6 +1,7 @@
 package emfswitch.operation;
 
 import emfswitch.SwitchImplementation;
+import java.lang.Double;
 import kmLogo.Equals;
 import kmLogo.Turtle;
 import org.eclipse.emf.ecoretools.ale.compiler.lib.EqualService;
@@ -18,11 +19,11 @@ public class EqualsOperation extends BinaryExpOperation {
 
 	public double eval(Turtle turtle) {
 		double result;
-		if(EqualService.equals((((ExpressionOperation) emfswitch.doSwitch(this.it.getLhs())).eval((Turtle) turtle)), (((ExpressionOperation) emfswitch.doSwitch(this.it.getRhs())).eval((Turtle) turtle)))) {
-			result = 1.0;
+		if(EqualService.equals((((ExpressionOperation) emfswitch.doSwitch(this.it.getLhs())).eval((Turtle) (turtle))), (((ExpressionOperation) emfswitch.doSwitch(this.it.getRhs())).eval((Turtle) (turtle))))) {
+			result = ((Double) (1.0));
 		}
 		else {
-			result = 0.0;
+			result = ((Double) (0.0));
 		}
 		return result;
 	}
