+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/switch/logo_lang/switch-comp/emfswitch/operation/IfOperation.java
@@ -1,6 +1,7 @@
 package emfswitch.operation;
 
 import emfswitch.SwitchImplementation;
+import java.lang.Double;
 import kmLogo.If;
 import kmLogo.Turtle;
 import org.eclipse.emf.ecoretools.ale.compiler.lib.EqualService;
@@ -18,15 +19,15 @@ public class IfOperation extends ControlStructureOperation {
 
 	public double eval(Turtle turtle) {
 		double result;
-		if(!EqualService.equals((((ExpressionOperation) emfswitch.doSwitch(this.it.getCondition())).eval((Turtle) turtle)), (0.0))) {
-			result = ((BlockOperation) emfswitch.doSwitch(this.it.getThenPart())).eval((Turtle) turtle);
+		if(!EqualService.equals((((ExpressionOperation) emfswitch.doSwitch(this.it.getCondition())).eval((Turtle) (turtle))), (0.0))) {
+			result = ((double) (((BlockOperation) emfswitch.doSwitch(this.it.getThenPart())).eval((Turtle) (turtle))));
 		}
 		else {
 			if(!EqualService.equals((this.it.getElsePart()), (null))) {
-				result = ((BlockOperation) emfswitch.doSwitch(this.it.getElsePart())).eval((Turtle) turtle);
+				result = ((double) (((BlockOperation) emfswitch.doSwitch(this.it.getElsePart())).eval((Turtle) (turtle))));
 			}
 			else {
-				result = 0.0;
+				result = ((Double) (0.0));
 			}
 		}
 		return result;
