+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/switch/logo_lang/switch-comp/emfswitch/operation/LowerOperation.java
@@ -1,6 +1,7 @@
 package emfswitch.operation;
 
 import emfswitch.SwitchImplementation;
+import java.lang.Double;
 import kmLogo.Lower;
 import kmLogo.Turtle;
 
@@ -17,11 +18,11 @@ public class LowerOperation extends BinaryExpOperation {
 
 	public double eval(Turtle turtle) {
 		double result;
-		if((((ExpressionOperation) emfswitch.doSwitch(this.it.getLhs())).eval((Turtle) turtle)) < (((ExpressionOperation) emfswitch.doSwitch(this.it.getRhs())).eval((Turtle) turtle))) {
-			result = 1.0;
+		if((((ExpressionOperation) emfswitch.doSwitch(this.it.getLhs())).eval((Turtle) (turtle))) < (((ExpressionOperation) emfswitch.doSwitch(this.it.getRhs())).eval((Turtle) (turtle)))) {
+			result = ((Double) (1.0));
 		}
 		else {
-			result = 0.0;
+			result = ((Double) (0.0));
 		}
 		return result;
 	}
