+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/switch/logo_lang/switch-comp/emfswitch/operation/RightOperation.java
@@ -18,9 +18,9 @@ public class RightOperation extends PrimitiveOperation {
 
 	public double eval(Turtle turtle) {
 		double result;
-		double angle = ((double) (((ExpressionOperation) emfswitch.doSwitch(this.it.getAngle())).eval((Turtle) turtle)));
+		double angle = ((double) (((ExpressionOperation) emfswitch.doSwitch(this.it.getAngle())).eval((Turtle) (turtle))));
 		((TurtleOperation) emfswitch.doSwitch(turtle)).rotate((Double) (-(angle)));
-		result = 0.0;
+		result = ((Double) (0.0));
 		return result;
 	}
 }
