+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/switch/testoperationbody/switch-comp/emfswitch/operation/MainOperation.java
@@ -87,7 +87,7 @@ public class MainOperation {
 		}
 		int cptr = ((int) (1));
 		while ((cptr) < (((MainOperation) emfswitch.doSwitch(this.it)).ma())) {
-			cptr = (cptr) + (1);
+			cptr = ((Integer) ((cptr) + (1)));
 		}
 		boolean truz = ((boolean) (!(false)));
 		boolean gt = ((boolean) ((1) > (2)));
@@ -103,7 +103,7 @@ public class MainOperation {
 
 	public int ma() {
 		int result;
-		result = 452;
+		result = ((Integer) (452));
 		return result;
 	}
 }
