+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/ConstantOperationImpl.java
@@ -18,7 +18,7 @@ public class ConstantOperationImpl extends ExpressionOperationImpl implements Co
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = this.it.getValue();
+		result = ((double) (this.it.getValue()));
 		return result;
 	}
 }
