+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/CosOperationImpl.java
@@ -20,7 +20,7 @@ public class CosOperationImpl extends UnaryExpressionOperationImpl implements Co
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = TrigoServices.cosinus((double) (((ExpressionOperation)this.it.getExpression().accept(vis)).eval((Turtle) (turtle))));
+		result = ((double) (TrigoServices.cosinus((double) (((ExpressionOperation)this.it.getExpression().accept(vis)).eval((Turtle) (turtle))))));
 		return result;
 	}
 }
