+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/LowerOperationImpl.java
@@ -1,5 +1,6 @@
 package visitor.operation.kmLogo.impl;
 
+import java.lang.Double;
 import kmLogo.visitor.kmLogo.Lower;
 import kmLogo.visitor.kmLogo.Turtle;
 import visitor.VisitorInterface;
@@ -20,10 +21,10 @@ public class LowerOperationImpl extends BinaryExpOperationImpl implements LowerO
 	public double eval(Turtle turtle) {
 		double result;
 		if((((ExpressionOperation)this.it.getLhs().accept(vis)).eval((Turtle) (turtle))) < (((ExpressionOperation)this.it.getRhs().accept(vis)).eval((Turtle) (turtle)))) {
-			result = 1.0;
+			result = ((Double) (1.0));
 		}
 		else {
-			result = 0.0;
+			result = ((Double) (0.0));
 		}
 		return result;
 	}
