+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/ProcCallOperationImpl.java
@@ -1,5 +1,7 @@
 package visitor.operation.kmLogo.impl;
 
+import java.lang.Double;
+import java.lang.Integer;
 import kmLogo.visitor.kmLogo.Expression;
 import kmLogo.visitor.kmLogo.KmLogoFactory;
 import kmLogo.visitor.kmLogo.ProcCall;
@@ -35,10 +37,10 @@ public class ProcCallOperationImpl extends ExpressionOperationImpl implements Pr
 			newVar.setName(CollectionService.get(this.it.getDeclaration().getArgs(), i).getName());
 			newVar.setValue(((ExpressionOperation)exp.accept(vis)).eval((Turtle) (turtle)));
 			newFrame.getVariables().add(newVar);
-			i = (i) + (1);
+			i = ((Integer) ((i) + (1)));
 		}
 		turtle.getCallStack().getFrames().add(newFrame);
-		result = 0.0;
+		result = ((Double) (0.0));
 		if(this.it.getDeclaration() instanceof ProcDeclaration) {
 			ProcDeclaration decl = ((ProcDeclaration) (this.it.getDeclaration()));
 			((ProcDeclarationOperation)decl.accept(vis)).deval((Turtle) (turtle));
