+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/ProcDeclarationOperationImpl.java
@@ -1,5 +1,6 @@
 package visitor.operation.kmLogo.impl;
 
+import java.lang.Double;
 import kmLogo.visitor.kmLogo.Instruction;
 import kmLogo.visitor.kmLogo.ProcDeclaration;
 import kmLogo.visitor.kmLogo.Turtle;
@@ -20,16 +21,16 @@ public class ProcDeclarationOperationImpl extends InstructionOperationImpl imple
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = 0.0;
+		result = ((Double) (0.0));
 		for(Instruction instr: this.it.getInstructions()) {
-			result = ((InstructionOperation)instr.accept(vis)).eval((Turtle) (turtle));
+			result = ((double) (((InstructionOperation)instr.accept(vis)).eval((Turtle) (turtle))));
 		}
 		return result;
 	}
 
 	public double deval(Turtle turtle) {
 		double result;
-		result = ((ProcDeclarationOperation)this.it.accept(vis)).eval((Turtle) (turtle));
+		result = ((double) (((ProcDeclarationOperation)this.it.accept(vis)).eval((Turtle) (turtle))));
 		return result;
 	}
 }
