+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/RightOperationImpl.java
@@ -23,7 +23,7 @@ public class RightOperationImpl extends PrimitiveOperationImpl implements RightO
 		double result;
 		double angle = ((double) (((ExpressionOperation)this.it.getAngle().accept(vis)).eval((Turtle) (turtle))));
 		((TurtleOperation)turtle.accept(vis)).rotate((Double) (-(angle)));
-		result = 0.0;
+		result = ((Double) (0.0));
 		return result;
 	}
 }
