+++ tests/org.eclipse.emf.ecoretools.ale.compiler.test/test-results/visitor/logo_lang/visitor-comp/visitor/operation/kmLogo/impl/TanOperationImpl.java
@@ -20,7 +20,7 @@ public class TanOperationImpl extends UnaryExpressionOperationImpl implements Ta
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = TrigoServices.tan((double) (((ExpressionOperation)this.it.getExpression().accept(vis)).eval((Turtle) (turtle))));
+		result = ((double) (TrigoServices.tan((double) (((ExpressionOperation)this.it.getExpression().accept(vis)).eval((Turtle) (turtle))))));
 		return result;
 	}
 }
